Created a for-loop thing.
The interesting question is about the fact that it's got no block to hold N statements, instead it's just got one statement at the end, possibly even null.

It very likely would be fine recursing, but I'm going to hold off on putting it in the <line> rule for the time being.